+++ bin/slack-messaging
@@ -16,7 +16,12 @@ program_long_desc """
 DOCUMENTATION
 """
 
-commands_from File.expand_path(File.dirname(__FILE__) + '/../lib/slack_messaging/commands')
+desc 'Prints a variety of messages to Slack'
+command 'slack' do |c|
+  c.action do |global_options, options, args|
+    SlackMessaging::Slack.execute(args, options) # args are optional string to print
+  end
+end
 
 pre do |global,command,options,args|
   SlackMessaging::Config.load(global[:config])
