+++ lib/slack_messaging.rb
@@ -1,9 +1,4 @@
-require_relative 'slack_messaging/version'
-require_relative 'slack_messaging/output'
-require_relative 'slack_messaging/config'
-require_relative 'slack_messaging/notify_slack'
-require_relative 'slack_messaging/random_message'
+require 'require_all'
+require_all 'lib/slack_messaging'
 
-module SlackMessaging
-
-end
+module SlackMessaging; end
